+++ ExactRiemannSolverBase.cpp
@@ -1203,7 +1203,7 @@ ExactRiemannSolverBase::ComputeRhoUStar(int wavenumber /*1 or 3*/,
 	    ps_0 = integrationPath[index0][0];
 	    rhos_0 = integrationPath[index0][1];
 	    us_0 = integrationPath[index0][2];
-	    if (index0 != integrationPath.size()-1) { // new starting point is not the last on the trajectory
+	    if (index0 != (int)integrationPath.size()-1) { // new starting point is not the last on the trajectory
               dp = ps_0-ps; 
             } else { // dp from the last step 
               dp = std::min( integrationPath[index0-1][0]-integrationPath[index0][0], ps_0-ps );
